Add DEBUG counter for internal Terminate calls to diagnostics
Tracks Terminate(false) calls to verify whether aborted states occur
but are cleared too fast for the 250ms diagnostic snapshots to capture.

+++ src/Listener.cpp
@@ -208,7 +208,7 @@ void Listener::Notify()
 
 void Listener::CheckPortCapacity()
 {
-	for (int i = 0; i < MaxConnections; i++)
+	for (size_t i = 0; i < MaxConnections; i++)
 	{
 		if (this != listeners[i])
 		{
